+++ packages/format-message/__tests__/react.spec.js
@@ -1,28 +1,28 @@
 /* eslint-env mocha */
-var expect = require('chai').expect
-var React = require('react')
-var formatChildren = require('../packages/format-message/react').formatChildren
+const expect = require('chai').expect
+const React = require('react')
+const formatChildren = require('../react').formatChildren
 
 describe('react formatChildren with numeric index tags', function () {
   it('returns a single child for simple messages', function () {
-    var results = formatChildren('simple')
+    const results = formatChildren('simple')
     expect(results).to.equal('simple')
   })
 
   it('preserves tokens with no element mapping', function () {
-    var results = formatChildren('<0>simple</0>')
+    const results = formatChildren('<0>simple</0>')
     expect(results).to.equal('<0>simple</0>')
   })
 
   it('returns a single child for wrapped messages', function () {
-    var results = formatChildren('<0>simple</0>', [
+    const results = formatChildren('<0>simple</0>', [
       React.createElement('span')
     ])
     expect(results).to.deep.equal(React.createElement('span', { key: '0' }, 'simple'))
   })
 
   it('preserves the props of the wrappers', function () {
-    var results = formatChildren('<0>simple</0>', [
+    const results = formatChildren('<0>simple</0>', [
       React.createElement('span', { className: 'foo' })
     ])
     expect(results).to.deep.equal(React.createElement('span', {
@@ -32,7 +32,7 @@ describe('react formatChildren with numeric index tags', function () {
   })
 
   it('returns an array of children when there are many', function () {
-    var results = formatChildren('it was <0>his</0> fault', [
+    const results = formatChildren('it was <0>his</0> fault', [
       React.createElement('em')
     ])
     expect(results).to.deep.equal([
@@ -43,7 +43,7 @@ describe('react formatChildren with numeric index tags', function () {
   })
 
   it('nests arbitrarily deep', function () {
-    var results = formatChildren('<0><1><2><3>deep</3></2></1></0>', [
+    const results = formatChildren('<0><1><2><3>deep</3></2></1></0>', [
       React.createElement('div'),
       React.createElement('span'),
       React.createElement('em'),
@@ -86,19 +86,19 @@ describe('react formatChildren with numeric index tags', function () {
 
 describe('react formatChildren with string tags', function () {
   it('preserves tokens with no element mapping', function () {
-    var results = formatChildren('<span>simple</span>')
+    const results = formatChildren('<span>simple</span>')
     expect(results).to.equal('<span>simple</span>')
   })
 
   it('returns a single child for wrapped messages', function () {
-    var results = formatChildren('<span>simple</span>', {
+    const results = formatChildren('<span>simple</span>', {
       span: React.createElement('span')
     })
     expect(results).to.deep.equal(React.createElement('span', { key: 'span' }, 'simple'))
   })
 
   it('preserves the props of the wrappers', function () {
-    var results = formatChildren('<span>simple</span>', {
+    const results = formatChildren('<span>simple</span>', {
       span: React.createElement('span', { className: 'foo' })
     })
     expect(results).to.deep.equal(React.createElement('span', {
@@ -108,7 +108,7 @@ describe('react formatChildren with string tags', function () {
   })
 
   it('returns an array of children when there are many', function () {
-    var results = formatChildren('it was <em>his</em> fault', {
+    const results = formatChildren('it was <em>his</em> fault', {
       em: React.createElement('em')
     })
     expect(results).to.deep.equal([
@@ -119,7 +119,7 @@ describe('react formatChildren with string tags', function () {
   })
 
   it('nests arbitrarily deep', function () {
-    var results = formatChildren('<div><span><em><strong>deep</strong></em></span></div>', {
+    const results = formatChildren('<div><span><em><strong>deep</strong></em></span></div>', {
       div: React.createElement('div'),
       span: React.createElement('span'),
       em: React.createElement('em'),
@@ -160,7 +160,7 @@ describe('react formatChildren with string tags', function () {
   })
 
   it('ignores invalid wrapper tags', function () {
-    var result = formatChildren('<sp an>test</sp an>', { span: 'span' })
+    let result = formatChildren('<sp an>test</sp an>', { span: 'span' })
     expect(result).to.equal('<sp an>test</sp an>')
 
     result = formatChildren('<sp\nan>test</sp\nan>', { span: 'span' })